Adjust timeout for non-retry calls
Give non-retry calls the full timeout seconds to finish and also
don't log that you're going to retry them when you are not.

(cherry picked from commit I51cbe3905b55f035718245b45bfc34536b8ce602)
Closes-bug: 1602208

Note: most of the patch has been included in other backport merge
conflict resolution, that's why so few is left of it.  But the piece is
critical to fix the above bug.

Change-Id: I5921df520d12eb2e89ccd21f7ff0b7cfe46b231d