Allow setting log files for apache vhost
This allows the setting of the error and access file logs, as well as
the access log format. This was done in a similar fashion as one can
configure these ones in the keystone wsgi manifest.

 Conflicts:
	manifests/wsgi/apache.pp
	spec/defines/openstacklib_wsgi_apache_spec.rb

Change-Id: I5e7d3588b7b3b106813d6d37b55aa812273d04d6
(cherry picked from commit 546eb8fb4142b797cf6c5618d4cc512fa8773d0a)