Adapt to new type validation in puppetlabs-apache
The puppetlabs-apache module is enforcing more strict data type
validation[1]. This change updates the default values to adapt to that
change.

Note that we still need to fix the unit tests as these are affected by
the change in default values. We can fix these once the new release is
made in the puppetlabs-apache module.

[1] https://github.com/puppetlabs/puppetlabs-apache/commit/f41251e336fa3e7c31c19968ccee74121cf71e47

Closes-Bug: #1983300
Change-Id: Ic6f65fdcb8953e0307d8a9d34cecb7aa1821a018

Cherry-picked from: 8334afc2a6967093e061cf605e087777a594ee03

Change-Id: I65fe2af8feee8326a4625e21a035c70b45086cd5