Stop testing underlying default in puppetlabs-apache
When passing undef(nil) to a defined resource type then the value is
just ignored and the default value in the defined resource type is
used. Because of this any change in default value in the dependent
module can directly break our unit tests if we strictly assert behavior
with undef passed.

This change removes a few validations which effectively assert defaults
in puppetlabs-apache. They updated a few default values in 8.0.0 and
that is causing multiple failures now.

Related-Bug: #1983300
Change-Id: I7bbe54972143bd51e41f73ccc838fd222788b08d

Cherry-picked from: ba19e6dbd30fcd931d78877d4e262fbd2e3d1da1

Change-Id: I8223dea07b32e3b4209ad60e61dd4b55334f6aad